requirements.txt
fastapi
uvicorn
pymongo
passlib[bcrypt]
python-jose
aio_pika
websockets 
httpx